WIP: Patient dashboard + fixed the fckn auth
