+++ src/components/ui/cards/NotificationCard.tsx
@@ -1,5 +1,5 @@
 
-import { XCircleIcon, HeartIcon, CheckCircleIcon } from '@heroicons/react/24/outline'
+import { CircleX, HeartIcon, CheckCheckIcon } from "lucide-react";
 
 
 interface NotificationCardProps {
@@ -17,7 +17,7 @@ const NotificationCard: React.FC<NotificationCardProps> = ({ statusNotification,
       return (
         <article className="card-notification card-notification--cancelled">
           <header className="card-notification__header">
-            <XCircleIcon className="size-6 text-warning-dark" /> <h6 className="font-semibold">¡Tu reserva ha sido cancelada!</h6>
+            <CircleX className="size-6 text-warning-dark" /> <h6 className="font-semibold">¡Tu reserva ha sido cancelada!</h6>
           </header>
           <section className="card-notification__body">
             <p>Su cita con el médico <span className="font-semibold">{notification.doctorName}</span> ha sido confirmada para el día <span className="font-semibold">{notification.day}</span> a las <span className="font-semibold">{notification.hour}</span>. Por favor, llegue 10 minutos antes de su cita y traiga cualquier documentación médica relevante.</p>
@@ -39,7 +39,7 @@ const NotificationCard: React.FC<NotificationCardProps> = ({ statusNotification,
       return (
         <article className="card-notification card-notification--accepted">
           <header className="card-notification__header">
-            <CheckCircleIcon className="size-6 text-success-dark" /> <h6 className="font-semibold">¡Cita confirmada!</h6>
+            <CheckCheckIcon className="size-6 text-success-dark" /> <h6 className="font-semibold">¡Cita confirmada!</h6>
           </header>
           <section className="card-notification__body">
             <p>Su solicitud con el médico <span className="font-semibold">{notification.doctorName}</span> para el día <span className="font-semibold">{notification.day}</span> a las <span className="font-semibold">{notification.hour}</span> ha sido recibida. Le enviaremos una confirmación una vez que el médico revise su disponibilidad.</p>
