+++ src/types/next-auth.d.ts
@@ -1,4 +1,4 @@
-import NextAuth from "next-auth";
+import "next-auth";
 import { JWT } from "next-auth/jwt";
 
 declare module "next-auth" {
@@ -7,19 +7,25 @@ declare module "next-auth" {
     last_name: string;
     email: string;
     access_token: string;
-    groups: number[];
+    groups: any;
   }
 }
 declare module "next-auth" {
   interface Session {
-    user: User;
+    user: JWT;
     access_token: string;
   }
 }
 
 declare module "next-auth/jwt" {
   interface JWT {
-    user: User;
+    user: {
+      first_name: string;
+      last_name: string;
+      email: string;
+      access_token: string;
+      groups: any;
+    };
     access_token: string;
   }
 }
